feat(npmrc): adding "update" option to update existing profiles
