+++ packages/npmrc/src/npmrc.ts
@@ -1,15 +1,13 @@
 #!/usr/bin/env node
 /* istanbul ignore file */
 
-import {
-	createProfile,
-	deleteProfile,
-	listProfiles,
-	switchProfile,
-} from "./utilities.js";
-
 import { config } from "./parse.js";
+import { createProfile } from "./utilities/createProfile.js";
+import { deleteProfile } from "./utilities/deleteProfile.js";
+import { listProfiles } from "./utilities/listProfiles.js";
 import path from "node:path";
+import { switchProfile } from "./utilities/switchProfile.js";
+import { updateProfile } from "./utilities/updateProfile.js";
 
 const HOME = process.env.HOME;
 const NPMRC_STORE = path.join(HOME, ".envtools/npmrcs");
@@ -22,6 +20,15 @@ if (flags.create && parameters !== undefined) {
 		storeConfig: NPMRC_STORE_CONFIG,
 		storeLocation: NPMRC_STORE,
 		profileName: parameters["0"],
+		homeLocation: HOME,
+	});
+	process.exit(exitFlag);
+}
+
+if (flags.update && parameters !== undefined) {
+	const exitFlag = await updateProfile({
+		storeConfig: NPMRC_STORE_CONFIG,
+		storeLocation: NPMRC_STORE,
 		homeLocation: HOME,
 	});
 	process.exit(exitFlag);
